Include authorizations to all methods

+++ 5 - API REST/routes/eventsRoutes.go
@@ -3,7 +3,6 @@ package routes
 import (
 	"fmt"
 	"gabriel/api/models"
-	"gabriel/api/utils"
 	"net/http"
 	"strconv"
 
@@ -23,29 +22,16 @@ func getEvents(context *gin.Context) {
 
 func createEvent(context *gin.Context) {
 
-	token := context.Request.Header.Get("Authorization")
-
-	if token == "" {
-		context.JSON(http.StatusUnauthorized, gin.H{"message": "Not authorized"})
-		return
-	}
-
-	err := utils.VerifyToken(token)
-
-	if err != nil {
-		context.JSON(http.StatusUnauthorized, gin.H{"message": "Not authorized"})
-		return
-	}
-
 	var event models.Event
-	err = context.ShouldBindJSON(&event)
+	err := context.ShouldBindJSON(&event)
 
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message": "Could not parse request data."})
 		return
 	}
 
-	event.UserID = 1
+	userID := context.GetInt64("userID")
+	event.UserID = userID
 
 	err = event.Save()
 
@@ -78,64 +64,75 @@ func getEvent(context *gin.Context) {
 }
 
 func updateEvent(context *gin.Context) {
-	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
 
+	// Obter o id indicado no requisição
+	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message": "Could not parse event ID."})
 		return
 	}
 
-	_, err = models.GetEventByID(eventId)
-
+	// Obter o event indicado pelo ID
+	event, err := models.GetEventByID(eventId)
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not fetch event."})
 		return
 	}
 
+	// Verificar se o usuário logado é o mesmo que criou o evento
+	userID := context.GetInt64("userID")
+	if event.UserID != userID {
+		context.JSON(http.StatusUnauthorized, gin.H{"message": "Not authorized to update event."})
+		return
+	}
+
+	// Ler o body da requisição e criar o evento atualizado
 	var updatedEvent models.Event
-
 	err = context.ShouldBindJSON(&updatedEvent)
-
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message": "Could not parse request data."})
 		return
 	}
-
 	updatedEvent.ID = eventId
 
+	// Atualizar evento
 	err = updatedEvent.Update()
-
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not update event."})
 		return
 	}
-
 	context.JSON(http.StatusOK, gin.H{"message": "Event updated successfully."})
 }
 
 func deleteEvent(context *gin.Context) {
 
+	// Obter o id indicado no requisição
 	eventId, err := strconv.ParseInt(context.Param("id"), 10, 64)
-
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"message": "Could not parse event ID."})
 		return
 	}
 
+	// Obter o event indicado pelo ID
 	event, err := models.GetEventByID(eventId)
-
 	if err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not find ID."})
 		return
 	}
 
-	err = event.DeleteEvent()
+	// Verificar se o usuário logado é o mesmo que criou o evento
+	userID := context.GetInt64("userID")
+	if event.UserID != userID {
+		context.JSON(http.StatusUnauthorized, gin.H{"message": "Not authorized to delete event."})
+		return
+	}
 
+	// Deletar evento do banco
+	err = event.DeleteEvent()
 	if err != nil {
 		fmt.Println(err)
 		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not delete event. "})
 		return
 	}
-
 	context.JSON(http.StatusOK, gin.H{"message": "Event deleted successfully."})
 }
